requirements.txt
beautifulsoup4>=4.12.3
feedparser>=6.0.11
numpy>=1.26.4
pandas>=2.2.2
plotly>=5.24.1
python-dateutil>=2.9.0.post0
requests>=2.32.3
scikit-learn>=1.5.1
streamlit>=1.39.0
transformers>=4.46.0
torch>=2.4.1

# Optional fallback for dynamic pages
selenium>=4.25.0
webdriver-manager>=4.0.2
